check if project file exists

+++ bin/templates/scripts/cordova/lib/prepare.js
@@ -331,10 +331,14 @@ function handleBuildSettings (platformConfig, locations, infoPlist) {
     if (wkWebViewOnly) {
         // update CordovaLib Xcode project, too
         var pbxPath = path.join(project.projectDir, 'CordovaLib', 'CordovaLib.xcodeproj', 'project.pbxproj');
-        var xcodeproj = xcode.project(pbxPath);
-        xcodeproj.parseSync();
-        xcodeproj.updateBuildProperty('WK_WEB_VIEW_ONLY', '1');
-        fs.writeFileSync(pbxPath, xcodeproj.writeSync());
+        if (fs.existsSync(pbxPath)) {
+            var xcodeproj = xcode.project(pbxPath);
+            xcodeproj.parseSync();
+            xcodeproj.updateBuildProperty('WK_WEB_VIEW_ONLY', '1');
+            fs.writeFileSync(pbxPath, xcodeproj.writeSync());
+        } else {
+            events.emit('error', 'Project file does not exist at ' + pbxPath);
+        }
     }
 
     return Q();
